backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
Flask-Redis==0.4.0
Flask-JWT-Extended==4.5.3
Werkzeug==2.3.7
PyJWT==2.8.0
opencv-python==4.8.1.78
face-recognition==1.3.0
deepface==0.0.79
ultralytics==8.0.196
numpy==1.24.3
Pillow==10.0.1
requests==2.31.0
pytz==2023.3
pymysql==1.1.0
redis==5.0.1
js-cookie==3.0.5
dayjs==1.11.10